test: Tear down WCF server after integration tests finish
When we run integration tests, we launch a WCF server within the
application. However we never dispose the server object after
integration tests finish. In order to dispose the server object, I
create a new endpoint called "CloseConnection" which will
close the WCF server. When integration tests finish, it will make a
request to this endpoint to shut down the WCF server. Meanwhile, I also
clean up and refactor code in multiple files.