test: Create customized logger to log warnings and errors on committed files
As we are using StyleCop Anyalyzer to lint C# code, we want to show
warnings and errors in the last committed files for every PR request. We
create a customized logger "SALogger" to log warnings and errors in last
committed files instead of all files in the whole project. After we get
the logs, we send it back to PR as comments. Meanwhile, we added a
RingloggerParser to parse log file (log.bin) to txt format which can be
readable.