tunnel: Add rudimentary captive portal detection
This doesn't check for DNS hijacking, and the actual captive portal
detection mechanism is rather weak, but it does lay the groundwork for
making connections that are exempt from the tunnel. Probably Microsoft's
server should not be used or hardcoded like that, either. This commit is
meant to be expanded upon in followups.

`CaptivePortalDetection.TestConnectivity()` should be called from the
broker process.