UI: removed auto-connection on initial server selection from the connection view.
Previously, when the user is unprotected, the client automatically initiates a connection when a server is selected, which isn't intuitive.  With this change, in the same scenario, selecting a server will return them to the main view, where the user must click the connection toggle to connect to the server.