ui: Add settings option for captive portal detection
With the new captive portal detection feature, we need a new settings options to allow users to specify whether they wish to be notified when a captive portal is detected.  There is a new view to manage notifications, where the captive portal settings checkbox will be housed.  Selecting this checkbox will alter the user's settings file in their local app data folder.