ui: Add post captive portal login notification
After the application detects a network with a captive portal, a Windows
notification is displayed, prompting the user to turn off the vpn and
log in to the captive portal.  Once the vpn is turned off, the
application detects when the user logs in to the detected captive portal
network, and displays a Windows notification (after an initial grace
period of 10s) that prompts the user to turn on the vpn again to secure
their network.