test: set up AWS ec2 to do smoke test
In order to do smoke test, we need to simulate the real production
environment which is in Windows 10. However the CircleCI is current
using Windows server 2019 for all windows job and Windows server 2019
only has IE 11 installed. IE 11 is not supported by FxA anymore and
cannot show user login web page properly. This is why we start using VM
to do smoke test.