test: fix the bug on UI tests
There are a few bugs on UI tests. One of them is that the pipeline is
not able to terminate the Cloudformation stack because Remove-CFNStack
command needs user to confirm before terminating. The others are tests
related such as maximized browser hide the UI of the vpn client and
notification chevron is not auto closed on vm.