test: fix the bug on firefox close
There is a bug which is caused by closing firefox browser. If the
browser has multiple tabs, then firefox will popup a window to let user
to confirm to close multi-tabs. It will fail our UI test since our
timeout is only 5 seconds. In order to pass, we extend the timeout for
browser session to 10 seconds. In the meantime, we refactor the code to
improve code quality.